backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.20
passlib[bcrypt]>=1.7.4
python-jose[cryptography]==3.3.0
alembic==1.14.0
better-exceptions==0.3.3
psycopg2-binary==2.9.10
gunicorn==23.0.0